Last touches

+++ _itoa.c
@@ -1,15 +1,9 @@
 #include "main.h"
-/**
- * _itoa - my function to print nums
- * @i: the num
- * @strout: the string to save the num
- * @base: either hexa or octa or dec
- * @Return: string with the int
- */
+
 char * _itoa(int i, char *strout, int base)
 {
   char *str = strout;
- long  int digit, sign = 0;
+  int digit, sign = 0;
   if (i < 0) {
     sign = 1;
     i *= -1;
